// packages/querius/src/types.ts
export const enum Mask {
  // Fundamental types
  UNDEFINED = 1 << 0,
  NULL = 1 << 1,
  NUMBER = 1 << 2,
  STRING = 1 << 3,
  BOOLEAN = 1 << 4,
  OBJECT = 1 << 5,
  ARRAY = 1 << 6,
  ID = 1 << 7,

  // Complex types
  RECORD = 1 << 8,
  UNION = 1 << 9,
  ENTITY = 1 << 10,

  // Flags
  HAS_SUB_ENTITY = 1 << 11,
  HAS_NUMBER_FORMAT = 1 << 12,
  HAS_STRING_FORMAT = 1 << 13,
  PARSE_RESULT = 1 << 14,
}
// ...
export type SimpleTypeDef =
  // Sets are constant values
  | Set<string | boolean | number>

  // Case-insensitive enum sets
  | CaseInsensitiveEnumSet<string | boolean | number>

  // Numbers are primitive type masks (potentially multiple masks combined)
  | Mask;

export type ComplexTypeDef =
  // Objects, arrays, records, unions, and parse results are definitions
  ObjectDef | EntityDef | ArrayDef | RecordDef | UnionDef | ParseResultDef;

export type InternalTypeDef = SimpleTypeDef | ComplexTypeDef;

export type InternalObjectFieldTypeDef = InternalTypeDef | string;

export type InternalObjectShape = Record<string, InternalObjectFieldTypeDef>;

export const ARRAY_KEY = Symbol('array');
export const RECORD_KEY = Symbol('record');

export interface UnionTypeDefs {
  [ARRAY_KEY]?: InternalTypeDef;
  [RECORD_KEY]?: InternalTypeDef;
  [key: string]: ObjectDef | EntityDef;
}

export interface BaseTypeDef {
  mask: Mask;
  shapeKey: number;
  typenameField: string;
  typenameValue: string;
  idField: string;
  subEntityPaths: undefined | string | string[];
  values: Set<string | boolean | number> | undefined;
}

export type EntityMethods = Record<string, (...args: any[]) => any>;
// ...
export interface EntityDef<T extends InternalObjectShape = InternalObjectShape, M extends EntityMethods = {}>
  extends BaseTypeDef {
  mask: Mask.ENTITY;
  shape: T;
}

export interface ObjectDef<T extends InternalObjectShape = InternalObjectShape> extends BaseTypeDef {
  mask: Mask.OBJECT;
  shape: T;
}

export interface ArrayDef<T extends InternalTypeDef = InternalTypeDef> extends BaseTypeDef {
  mask: Mask.ARRAY;
  shape: T;
}

export interface UnionDef<_T extends readonly InternalTypeDef[] = readonly InternalTypeDef[]> extends BaseTypeDef {
  mask: Mask.UNION;
  shape: UnionTypeDefs | undefined;
}

export interface RecordDef<T extends InternalTypeDef = InternalTypeDef> extends BaseTypeDef {
  mask: Mask.RECORD;
  shape: T;
}

export interface ParseResultDef<T extends InternalTypeDef = InternalTypeDef> extends BaseTypeDef {
  mask: Mask.PARSE_RESULT;
  shape: T;
}
// ...
export interface MutationResultValue<Response> {
  readonly response: Response;
}